[Silver V] Title: 귀찮음, Time: 52 ms, Memory: 5224 KB, Score: 30 point -BaekjoonHub

+++ 백준/Silver/16208. 귀찮음/귀찮음.cc
@@ -14,7 +14,7 @@ int main() {
     FastIO();
 
     vector<int> v;
-    int n, answer = 0, sum = 0;
+    long long n, answer = 0, sum = 0;
     cin >> n;
     for(int i = n ; i >= 1 ; i--){
         int input;
